+++ example/lib/main.dart
@@ -1,26 +1,37 @@
 import 'package:flutter/material.dart';
 import 'package:xwidget/xwidget.dart';
-
-import 'xwidget/generated/controllers.g.dart';
-import 'xwidget/generated/icons.g.dart';
-import 'xwidget/generated/inflaters.g.dart';
+import 'package:xwidget_example/xwidget/generated/registry.g.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
 
+  // Local resources, no analytics (default)
+  await XWidget.initialize(register: registerXWidgetComponents);
+
   // Cloud settings should be loaded from Firebase's Remote Config
   // or similar service.
-  await XWidget.initialize(
-    // projectKey: '<your-project-key>',
-    // storageKey: '<your-storage-key>',
-    // channel: 'dev',
-    // version: '1.0.0',
-  );
 
-  // register XWidget components
-  registerXWidgetIcons();
-  registerXWidgetInflaters();
-  registerXWidgetControllers();
+  // Cloud resources and analytics
+  // See https://docs.xwidget.dev/latest/runtime/cloud_resources/
+  // await XWidget.initialize(
+  //     register: registerXWidgetComponents,
+  //     resources: CloudResources(
+  //         projectKey: "<projectKey>",
+  //         storageKey: "<storageKey>",
+  //         channel: "<channel>",
+  //         version: "<version>"
+  //     )
+  // );
+
+  // Local resources with cloud analytics
+  // See https://docs.xwidget.dev/latest/runtime/local_resources/
+  // await XWidget.initialize(
+  //     register: registerXWidgetComponents,
+  //     resources: LocalResources.withAnalytics(
+  //         projectKey: "<projectKey>",
+  //         version: "<version>"
+  //     )
+  // );
 
   runApp(const MyApp());
 }