+++ lib/xwidget.dart
@@ -8,7 +8,7 @@ export 'src/custom/media.dart';
 export 'src/custom/value_listener.dart';
 export 'src/utils/events.dart';
 export 'src/utils/parsers_material.dart';
-export 'src/utils/resources.dart';
+export 'src/utils/resources/resources.dart';
 export 'src/utils/utils.dart';
 export 'src/defaults.dart';
 export 'src/xwidget.dart';